rename: MyGoal 페이지를 MyPage로 변경
